Add Storage to AddServiceParams
Add support for storage constraints to bundle changes.
This will be used by juju/juju to deploy bundles with
services that require storage.

Storage constraints are being added to GUI arguments;
I'm not sure if this will have any negative effects,
or if extraneous arguments are simply ignored. Please
advise.