Detect inconsistent bundle-model machine mappings
... and return an error that steers towards the change to fix it.

When a user redeploys a bundle into a model where a previous deploy had
some failures which were then corrected manually, and they're using
--map-machines=existing, they can get an inconsistency between the
bundle and model machines such that bundlechanges creates a new machine
but then doesn't host any units on it. This causes a panic in `juju
deploy`, as seen in https://pad.lv/1773357.

For example, if you have this bundle:

        applications:
            memcached:
                charm: cs:xenial/memcached-21
                num_units: 2
                to: [0, 1]
        machines:
            0:
            1:

When you deployed it machine 0 failed, so you manually added another
unit of memcached (which created machine 2). If you then tried to redeploy
the bundle with `--map-machines=existing`, it would see that machine 0
doesn't exist, add a change to create it, see also that there are already
two units of memcached, so not create any more.

The way to resolve this is to add an explicit mapping that bundle
machine 0 corresponds to model machine 2
`--map-machines=existing,0=2`. So now we show an error asking for this
mapping - in this case it's pretty obvious, but the situation is more
likely to be the one in the bug above.