Convert bundle machine names via the model machine map
This means that instead of combining all of the machine IDs first and
then looping over the set, we instead collect up all of the model
machine IDs and cross them off as we go through the bundle
machines (mapping machine IDs as we go). Any model machines we didn't
cross off are then missing from the bundle.