Move params and Args() to each type class
The commit moves the params back to each type class, along with
moving the Args() method. The reason for this change is that the
handlers that walk through the types edits the types from the
outside and the params you expect to be created and edited up
front aren't the same ones that are modified later on. So in the
end, only the params of each type class get's changed, not the
ones in changeInfo. Which ultimately means that when it comes
to serialise the results, it ends up being wrong as data is
ends up being missed.

If this library ever get's re-written then a vistor pattern would
do wonders for encaspulation and handling of one's own
responsibilities. In the mean time, the following changes are a
step in the right direction.