Charm and charmrepo dependencies
As the charm brings in dependency breaking changes (prefix for bundles
are required), then we need to take that into account when doing
charmbundle changes.

The amount of changes inside go mod and sum are a bit concerning, but
all the tests pass, so I'm inclined to leave it as is for now.