Re-write the descriptions
The following updates the descriptions for all charmhub changes. It
drops the URL schema and revision in favour of plain text.

This makes it easier to the user about what is happening and when
without the understanding or parsing for the given output.

Note: this is a version bump to v4 to prevent v3 users breaking.