Upload application with arch/series constraint
So if an application has existing architecture or series information
then we should set the constraints for the given application. If a
bundle has constraints and series set then the application should upload
a new charm and register a change.

As each charm is a unique blob we have no way to say grab be all of
these series of things, instead we have to upload one at a time
potentially stopping us from getting what we want if one differs
(heterogeneous applications).

The code is rather simple as we've done all the heavy lifting, the
change is to ensure that the key for the application is unique for every
architecture and series.