Handle bundle series
When applying a bundle diff to a model and a bundle we should also take
into account the bundle series. The bundle series should be taken into
account as that is the same mechanism as Juju takes into account and how
bundle themselves parsing works.

This should make it much more efficient when comparing existing bundles
when series fields are empty and global series is filled.