Add application channel
Juju needs to know more about channels that an application can be given
and because of the way addCharm and addApplication is split it becomes
harder to understand the key change.

The code is simple, expose channel in the same way that series is on the
application and push it through the code.

There already exists a test for channels, so modifying that to show the
right thing was easy.