Upgrade charms based on channels
The following commit allows the upgrading of a charm based on the
channel. It works by checking if an existing application has the same
channel, series and architecture via a RevisionGetter as the supplied
bundle application.

The revision getter pages out to the consumer of the library via a
function that the library consumer can supply. The intent is to supply a
revision based on what is supplied and if that revision doesn't match
the existing one, then an upgrade is required.

The code isn't overly complicated, but triggering the tests from the
changes side isn't easy. To back that up, the supply of a new test suite
for the resolve (handlers.go) type helps iron out any bugs.