Pass channel back when resolving.
The following changes updates the revision getter to now pass back the
channel. Updating any code to ensure that we rename any old naming to
the new one.

Unfortunately we have to request the channel before we can check the
equality, previously we checked just the channels. This became a problem
when the existing application had been resolved even if the channel was
empty, yet the bundle was still shown to be empty. Effectively
preventing the deployment of bundles with empty channels. The new unit
tests ensure we get the correct error message based on the content of
the model and the bundle.